refactor: adjust data document name
