+++ src/components/element/IndexKV.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import { ref, onMounted, reactive, toRefs } from 'vue';
-import indexData from '../../data/indexData.json';
+import indexData from '../../data/indexPageData.json';
 
 const { candidateLinks } = indexData.components.IndexKV;
 const { candidateKVImgTsai } = toRefs(reactive({ candidateKVImgTsai: null }));
